+++ app/lib/analyzer/pana_runner.dart
@@ -156,10 +156,10 @@ class AnalyzerJobProcessor extends JobProcessor {
 
   Future<void> _storeScoreCard(Job job, Summary summary,
       {List<String> flags}) async {
-    await scoreCardBackend.updateReportAndCard(
+    await scoreCardBackend.updateReportOnCard(
       job.packageName,
       job.packageVersion,
-      panaReportFromSummary(summary, flags: flags),
+      panaReport: panaReportFromSummary(summary, flags: flags),
     );
   }
 }