+++ app/lib/shared/versions.dart
@@ -21,9 +21,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.06.01', // The current [runtimeVersion].
+  '2021.06.03', // The current [runtimeVersion].
+  '2021.06.01',
   '2021.05.25',
-  '2021.05.19',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,